backend/requirements.txt
agno>=2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
astronomy-engine>=2.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
